+++ changelog.md
@@ -1,6 +1,8 @@
 ## unreleased
+
 - Fix problems with OCI environment settings passed to kiddie-pool for runners.
 - Issue error message when store-oci and storepath are both specified in the wercker command
+- Fix issue with docker:true and after-steps (#538)
 
 ## v1.0.1401 (2018-11-02)